Frontend/i18n: Move date i18n utils to i18n folder
