+++ app/web/features/communities/events/LongEventCard.tsx
@@ -11,15 +11,12 @@ import { eventImagePlaceholderUrl } from "appConstants";
 import Pill from "components/Pill";
 import FlagButton from "features/FlagButton";
 import { useTranslation } from "i18n";
+import { BROWSER_TIMEZONE, localizeDateTimeRange } from "i18n/dates";
 import { COMMUNITIES } from "i18n/namespaces";
 import Link from "next/link";
 import { Event } from "proto/events_pb";
 import { routeToEvent } from "routes";
-import {
-  BROWSER_TIMEZONE,
-  localizeDateTimeRange,
-  timestamp2Date,
-} from "utils/date";
+import { timestamp2Date } from "utils/date";
 import dayjs from "utils/dayjs";
 
 const StyledCard = styled(Card)(({ theme }) => ({
